requirements.txt
flask
flask-cors
graphene
ftswarm-py==1.2.1
coloredlogs==15.0.1
colorama==0.4.6
pyserial==3.5
semantic_version==2.8.5
requests==2.26.0
tabulate==0.9.0
bottle==0.12.25
ajsonrpc==1.2.0
marshmallow==3.19.0
python-dotenv==1.0.0
python-statemachine[diagrams]==2.0.0
pydot
grpcio
grpcio-tools
grpcio-reflection